Avoid overflow on waitUntilConnected().

+++ src/main/java/org/I0Itec/zkclient/ZkClient.java
@@ -486,7 +486,7 @@ public class ZkClient implements Watcher {
     }
 
     public void waitUntilConnected() {
-        waitUntilConnected(Long.MAX_VALUE, TimeUnit.MILLISECONDS);
+        waitUntilConnected(Integer.MAX_VALUE, TimeUnit.MILLISECONDS);
     }
 
     public boolean waitUntilConnected(long time, TimeUnit timeUnit) {
